arm: move flush_dcache_all() to just before disable cache
In Cortex-A15 architecture, when we run cache invalidate
the cache clean operation executes automatically.
So if there are any dirty cache lines before disabling the L2 cache
these will be synchronized with the main memory when
invalidate_dcache_all() runs in the last part of U-boot

The two functions after flush_dcache_all is using the stack. So this
data will be on the cache. After disable when invalidate is called the
data will be flushed from cache to memory. This corrupts the stack in
invalida_dcache_all. So this change is required to avoid the u-boot
hang.

So flush has to be done just before clearing CR_C bit

BUG=chrome-os-partner:9388
TEST=Boot from sdcard on Daisy 2
Booted kernel 3 times till login screen

Change-Id: I7aa9eede3ecd8d9a61f4c5f61bdf1d76f23bd732
Signed-off-by: Arun Mankuzhi <arun.m@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/22686
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Prashanth Godrehal <prashanth.g@samsung.com>
Commit-Ready: Alim Akhtar <alim.akhtar@samsung.com>