x86: Don't expect coreboot to prepare an FDT for u-boot
libfdt inclusion was not accepted in upstream coreboot, because
all it was used for was preparing a device tree for u-boot, fixing
it up with information that coreboot already exports in the coreboot
table. Since u-boot already parses the coreboot table anyways, it
is the cleaner approach to include the device tree preparation in
u-boot's coreboot specific init code itself.

This change required:
 - Drop the code that finds an FDT pointer in the coreboot table.
 - Since the memory allocator is not initialized yet by the time
   cpu_init_f() is run, use a piece of (unused) CBMEM to store the FDT.

There is some more cleanup that could be done, like moving coreboot
specific code out of board/chromebook-x86 into arch/x86/cpu/coreboot
but that's a task for another day.

Signed-off-by: Stefan Reinauer <reinauer@chromium.org>
BUG=chrome-os-partner:9648
TEST=build with our new coreboot branch chromeos-2012.05.10 and
     observe u-boot being able to boot into ChromeOS (more CLs
     needed though)

Change-Id: I2effc637f991bdb5c1b00b7327693de27702db8f
Reviewed-on: https://gerrit.chromium.org/gerrit/23275
Tested-by: Stefan Reinauer <reinauer@chromium.org>
Reviewed-by: Ronald G. Minnich <rminnich@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>