exynos: edid: Check the EDID on the FDT when bootup
When u-boot bootup, it checks the EDID info on the FDT correct or not.
If it is not available, try to probe the display device to get one.

This EDID data is needed in later u-boot and kernel for configuration.

BUG=chrome-os-partner:9449
TEST=Built with DEBUG defined and booted u-boot to see the edid messages show.
Removed the edid field in FDT and did the above steps to see the messages too.

Change-Id: Iddf0aeee64d2627c86e38a7cc2aca7624f4a9dd7
Signed-off-by: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/23509