exynos5: move the system_clock_init after the power_init being called
We were have problems with software reboot from the kernel were it was
not reliable, most likely the problem happens to be at the time of
sofware reset (lower operating point) voltage for the CPU has been
lowered.  After reset though the cpu speed is brought up but the
voltages are not.  This CL calls power_init() before calling
system_clock_init to make sure that voltages are brought up before
setting the clock.

We still need to see if i2c works properly as PMIC need to programed
over i2c.

BUG=chrome-os-partner:9325
TEST=build ok, booted on snow, booted kernel, soft reboot works 10 out
of 10 times

Change-Id: I63fedd85effa3c9fb3d4d897b9abae1a5fdec5e3
Signed-off-by: Arjun.K.V <arjun.kv@samsung.com>
Signed-off-by: Alim Akhtar <alim.akhtar@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/23552
Reviewed-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>
Commit-Ready: Doug Anderson <dianders@chromium.org>