power: Add generic battery charge controller
This supports a battery system comprising a TI TPS65090 PMIC / battery
charger, and a smart battery (Smart Battery Specification). A standard
interface is provided which other drivers could be  plugged into if
required.

BUG=chrome-os-partner:9028
TEST=manual:
build and boot on daisy with battery
'battery loop' and see that it communicates ok and charges the battery

Change-Id: Ic39e0245c82efcce7cb8d280df860295c6303f26
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/23786