cros: Rename file name prefix from cros_gpio to vboot_flag
While ARM platform read hardware flags from GPIO, x86 platforms read
hardware flags from sysinfo, which is a table provided by coreboot; but
due to some historical design issues of the API, sysinfo has pretended
to be GPIO, which is quite odd.

This is a series of patches that refactors and merges the two types of
hardware flags into a single, unified API.  And it is designed to be
extensible to new type of hardware flags as well.

BUG=chrome-os-partner:9706,9183
TEST=On Daisy and Link U-Boot command prompt, run
         vbexport_test init ; vboot_test flag
     And observe correct flag values.  Then boot to CrOS and run
         crossystem
     And observe correct output.

Change-Id: I78c28c6b3e862d3715b2c711479733803057ee80
Signed-off-by: Che-Liang Chiou <clchiou@chromium.org>
Signed-off-by: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/24861