x86: Drop unused entries from coreboot data structure
libfdt inclusion was not accepted in upstream coreboot, because
all it was used for was preparing a device tree for u-boot, fixing
it up with information that coreboot already exports in the coreboot
table. Since u-boot already parses the coreboot table anyways, it
is the cleaner approach to include the device tree preparation in
u-boot's coreboot specific init code itself.

This change required:
 - Drop unused entries in coreboot specific struct sysinfo_t.

Signed-off-by: Stefan Reinauer <reinauer@chromium.org>
BUG=chrome-os-partner:9648
TEST=build with our new coreboot branch chromeos-2012.05.10 and
     observe u-boot being able to boot into ChromeOS

Change-Id: Iff4efaf0f7ab1903989c69c04efaa97a7def269e
Reviewed-on: https://gerrit.chromium.org/gerrit/23930
Tested-by: Stefan Reinauer <reinauer@chromium.org>
Reviewed-by: Ronald G. Minnich <rminnich@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>