x86: Add coreboot serial port information to FDT
libfdt inclusion was not accepted in upstream coreboot, because
all it was used for was preparing a device tree for u-boot, fixing
it up with information that coreboot already exports in the coreboot
table. Since u-boot already parses the coreboot table anyways, it
is the cleaner approach to include the device tree preparation in
u-boot's coreboot specific init code itself.

This change required:
 - Make coreboot table parsing code for serial ports more generic
   and simpler.
 - Add a function to add serial console information from the coreboot
   table to the FDT.

Signed-off-by: Stefan Reinauer <reinauer@chromium.org>
BUG=chrome-os-partner:9648
TEST=build with our new coreboot branch chromeos-2012.05.10 and
     observe u-boot being able to boot into ChromeOS (more CLs
     needed)

Change-Id: Ie92ba6119bfbee5c9efe1aa2f00e67a747030192
Reviewed-on: https://gerrit.chromium.org/gerrit/23931
Tested-by: Stefan Reinauer <reinauer@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>