cros: Add ChromeOS specific data pointer to FDT
libfdt inclusion was not accepted in upstream coreboot, because
all it was used for was preparing a device tree for u-boot, fixing
it up with information that coreboot already exports in the coreboot
table. Since u-boot already parses the coreboot table anyways, it
is the cleaner approach to include the device tree preparation in
u-boot's coreboot specific init code itself.

This change required:
 - Add chromeos.[ch], implementing code to find the ChromeOS specific
   vdat area in the coreboot table and add it to the FDT. This code
   is taken mostly unmodified from existing coreboot code

Signed-off-by: Stefan Reinauer <reinauer@chromium.org>
BUG=chrome-os-partner:9648
TEST=build with our new coreboot branch chromeos-2012.05.10 and
     observe u-boot being able to boot into ChromeOS

Change-Id: I2d8105b56ab6f333677f0ff0c2497f7e20093796
Reviewed-on: https://gerrit.chromium.org/gerrit/23932
Tested-by: Stefan Reinauer <reinauer@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>