fdt: Add coreboot serial port information to FDT
libfdt inclusion was not accepted in upstream coreboot, because
all it was used for was preparing a device tree for u-boot, fixing
it up with information that coreboot already exports in the coreboot
table. Since u-boot already parses the coreboot table anyways, it
is the cleaner approach to include the device tree preparation in
u-boot's coreboot specific init code itself.

This change required:
 - Add a serial port node to each supported mainboard, as requested by
   Simon.

Signed-off-by: Stefan Reinauer <reinauer@chromium.org>
BUG=chrome-os-partner:9648
TEST=build with our new coreboot branch chromeos-2012.05.10 and
     observe u-boot being able to boot into ChromeOS (more CLs
     needed)

Change-Id: Icf1ff3b598e050473f6b33caaab2bfc8a7089508
Reviewed-on: https://gerrit.chromium.org/gerrit/24045
Tested-by: Stefan Reinauer <reinauer@chromium.org>
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>