gen: mkbp: Add support for the mkbp command to get information about the matrix
This change adds support for the command which asks the EC for information
about the keyboard matrix. The structure that's returned has a field called
"switches" which holds a bit vector of switches the EC has detected, including
the keyboard based recovery switch.

It also adds a subcommand to the mkbp command called "info" which prints out
the structure returned by the EC.

BUG=chrome-os-partner:8974
TEST=Built and booted on Daisy and ran the command "mkbp info". Verified that
the EC received the correct command and that the printed values matched what
the EC sent. With some minor adjustments to the EC provided by vpalatin, used
keyboard based recovery on my Daisy with a Mario keyboard and verified that a
bit in the "switches" value toggled.

Signed-off-by: Gabe Black <gabeblack@google.com>

Change-Id: I63e937787618675a47160d8afa78132dab6015f3
Reviewed-on: https://gerrit.chromium.org/gerrit/24181
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Ready: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>