drop: exynos: Hack to quiet MIPI errors in ChromeOS
This change is a terrible hack which somehow makes a class of MIPI related
errors on the console go away once booted into ChromeOS. It is not a long term
solution, and also breaks the LCD console in U-Boot.

BUG=chrome-os-partner:9997
TEST=Built and booted into ChromeOS and saw the MIPI DSI errors where the
intsrc register was 0 go away.

Change-Id: I2464c9086b64f94b1eaac61c5ce46e558c6e551c
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/24095
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Ready: Gabe Black <gabeblack@chromium.org>