drop: arm: exynos: Disable the display controller when starting Linux
If the display controller is still active when Linux starts, it may try to
access the frame buffer when the sysmmu is enabled. That causes the kernel to
panic. This change disables the display controller by unsetting the last two
register values its driver sets in the opposite order than they were set in.
It was determined experimentally that setting these two registers prevent the
kernel panic mentioned above, but also avoid putting the hardware into a state
where the kernel isn't able to set up graphics properly afterwards.

BUG=chrome-os-partner:10017
TEST=Built and booted on a kernel with sysmmu and saw the kernel panic go away.
Booted on the ToT kernel and saw that graphics still worked.

Change-Id: I1eff720d734006ecc8ee4a5a0cef9d1d1d400c11
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/24178
Commit-Ready: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>