exynos: spl: Store gd struct in the stack for SPL serial init
This CL is for preparing moving serial init earlier.

The gd struct is only needed for serial initialization. Since the
early_serial_init() is called in SPL u-boot. We store the gd struct in the
stack instead of the default memory region which may not be initialized.

BUG=chrome-os-partner:9961
TEST=Built and booted u-boot. Still see the SPL message:
U-Boot SPL, board rev 4

Change-Id: I1bd63484935031840389bbbb3207179892b86fde
Signed-off-by: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/24296
Reviewed-by: Simon Glass <sjg@chromium.org>