cros: Fix compiler warning of not used cros_gpio_dump
When VBOOT_DEBUG is not defined, cros_gpio_dump is defined as a static
empty function but not actually used in some modules, and compilers
warns you about this.  So remove the static modifier to fix the warning.

Signed-off-by: Che-Liang Chiou <clchiou@chromium.org>

BUG=chrome-os-partner:10144
TEST=run `USE=-cros-debug emerge-daisy chromeos-u-boot` and see no
     warnings on:

error: 'cros_gpio_dump' defined but not used [-Werror=unused-function]

Change-Id: I2dea15b521e8c2a4a5f18fbcd333e353f07f8402
Reviewed-on: https://gerrit.chromium.org/gerrit/24384
Commit-Ready: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>