gen: usb: Fix some incorrect cache/memory handling in the EHCI driver
The EHCI driver was not flushing the right areas of memory when setting up
structures which were consumed by the host controller. Specifically, it wasn't
flushing out all of the transaction descriptor structures, only the last one
it had added. Also, the driver didn't always use functions which guarantee
that writes actually make it to memory when they need to, or that they're
ordered properly relative to flushing operations. These problems were causing
the driver not to work properly when caches were enabled.

Many of the problems mentioned above are corrected in this change. Not all of
them were strictly necessary to get it to work with caching turned on, but it
should make the code more robust and correct and less likely to break in the
future.

It would be worthwhile to investigate whether the ehci_readl and ehci_writel
functions should be replaced with the normal readl and writel functions. The
former use volatile pointers to ensure that values are actually written out,
but they preclude an architecture specific implementation from adding
read/write barriers which might be more correct.

Also, this change adds additional padding to the QH structure to ensure that
its size is a multiple of 64 bytes. This isn't an ideal solution because it
only ensures that it's a multiple of 32 and 64 bytes, not any larger cache line
size, and is slightly wasteful if the cache line size really is 32 bytes.

BUG=chrome-os-partner:10215
TEST=Built with caches enabled without this change and saw errors related to
alignment when running the "usb start" command. Fixed those and saw errors
related to memory handling (the earlier errors scrolled the framebuffer
console, effectively clearing the cache). Implemented the remainder of this
change and saw "usb start" complete correctly. Booted ChromeOS from a USB key.
Built for Link.

Signed-off-by: Gabe Black <gabeblack@google.com>

Change-Id: I81fa88c7be91f8f158496db8fa7e4a38b2df30cd
Reviewed-on: https://gerrit.chromium.org/gerrit/25074
Commit-Ready: Gabe Black (Do Not Use) <gabeblack@google.com>
Reviewed-by: Gabe Black (Do Not Use) <gabeblack@google.com>
Tested-by: Gabe Black (Do Not Use) <gabeblack@google.com>