exynos: spi: Rework the driver so that it's more reliable
This change reworks the driver so that it's a bit simpler and more reliable.
That's done in several ways.

The spi_flush_fifo function doesn't have to manually empty the RX FIFO since
the software reset it does shortly after does it automatically. That code is
removed.

The driver is now a single loop which fills the TX FIFO as space is available
and reads from the RX FIFO as data is available. The RX and TX channels are
left on all the time. The controller is software reset only at the start of a
transaction.

The driver no longer tries to control the CS line manually through the
controller and also manually by turning it into a GPIO line. The controller is
now responsible for that line.

BUG=chrome-os-partner:10186
BUG=chrome-os-partner:10348
TEST=Used the driver with the speed in the device tree set to 5000000. The
initial probe to identify the flash chip and then reads to read in the
environment succeeded. sf probe 1:0 succeeded. sspi 1:0 128 9f succeeded.
saving the environment succeeded which erases flash, polls the status register,
sends a write enable, and writes new pages.

Signed-off-by: Gabe Black <gabeblack@google.com>

Change-Id: Ibb0edcc5b0c4e2431eb8518c1e949dba47e2491d
Reviewed-on: https://gerrit.chromium.org/gerrit/25282
Reviewed-by: Gabe Black (Do Not Use) <gabeblack@google.com>
Tested-by: Gabe Black (Do Not Use) <gabeblack@google.com>
Commit-Ready: Gabe Black <gabeblack@chromium.org>