x86: cros: split: Fix data corruption of ChromeOS ACPI data
coreboot used to pass some information to u-boot in the coreboot table
and other information in a modified flat device tree. Since the FDT code
was never upstreamed and removed from our tree, u-boot was changed to
get the information it needs from the coreboot table alone. However,
in the process of this change only the vboot shared data structure was
passed on by coreboot, so when u-boot tried to update the ChromeOS
specific ACPI entries, it would accidently overwrite the vboot data.

After the original switch to the new upstream coreboot, u-boot's init
code would parse the coreboot table for the ChromeOS ACPI data pointer,
add that to the device tree, and read it back from the device tree in
in an inconsistent way. Instead of doing this, use the value from the
coreboot table directly, and make sure we use the ChromeOS ACPI data
when updating ACPI values, and the vboot shared data when passing a
buffer to vboot.

Signed-off-by: Stefan Reinauer <reinauer@chromium.org>
BUG=chrome-os-partner:10252
TEST=Run crossystem |grep devsw_boot and observe it no longer prints
     error:
       devsw_boot             = 1 # Developer switch position at boot

Change-Id: Ib62aa4b54e09a14fde938f543fab84a77c1da790
Reviewed-on: https://gerrit.chromium.org/gerrit/25238
Tested-by: Stefan Reinauer <reinauer@google.com>
Reviewed-by: Ronald G. Minnich <rminnich@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>