cros: fdt: drop unused entry from device tree
coreboot used to pass some information to u-boot in the coreboot table
and other information in a modified flat device tree. Since the FDT code
was never upstreamed and removed from our tree, u-boot was changed to
get the information it needs from the coreboot table alone. However,
in the process of this change only the vboot shared data structure was
passed on by coreboot, so when u-boot tried to update the ChromeOS
specific ACPI entries, it would accidently overwrite the vboot data.

Previous patches fixed this issue. This one removes the remnants of the
old code from chromeos-coreboot.dtsi

Signed-off-by: Stefan Reinauer <reinauer@chromium.org>
BUG=chrome-os-partner:10252
TEST=No functional change, just removed dead code

Change-Id: I53847c147a5ac69e11d9acd4e4f36b8df453a08d
Reviewed-on: https://gerrit.chromium.org/gerrit/25239
Tested-by: Stefan Reinauer <reinauer@google.com>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Commit-Ready: Stefan Reinauer <reinauer@chromium.org>