spi: Fix spi_flash_cmd_write_multi so it calculates page_size properly
The spi_flash_cmd_write command was calculating the size of a page of flash by
shifting 1 to the left by the size of the page instead of just using the page
directly. This was the result of a mistake when rebasing our changes. Because
the page size is likely larger than the width of the constant 1 in bits, the
exact behavior is undefined. What's likely to happen, however, is that the
shift amount will be truncated. That will likely result in no shift at all, and
a page_size of 1. That will probably still work, but will be very inefficient.

BUG=chrome-os-partner:10186
TEST=Printed the computed page size and the write transactions used to write
the flash and saw correct behavior with this change and incorrect behavior
without it.

Signed-off-by: Gabe Black <gabeblack@google.com>

Change-Id: I0109141f21f25c5a113a414361d156cf7129b3e2
Reviewed-on: https://gerrit.chromium.org/gerrit/25284
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>