exynos5: pinumx: Simplify the SPI pinmux code
Now that the SPI CS line is handled by the SPI controller itself, there's no
need for the pinmux code to set it up as a GPIO. This change removes the
now unnecessary code associated with that feature, and simplifies the code
that remains.

BUG=chrome-os-partner:10186
BUG=chrome-os-partner:10348
TEST=Built and booted on Daisy which verifies that probes and reads work. Used
the "save" command and rebooted which verifies that writes work.

Signed-off-by: Gabe Black <gabeblack@google.com>

Change-Id: I5df94b4eca5f32b9a2d1eb8e76e392421178710b
Reviewed-on: https://gerrit.chromium.org/gerrit/25283
Reviewed-by: Gabe Black (Do Not Use) <gabeblack@google.com>
Tested-by: Gabe Black (Do Not Use) <gabeblack@google.com>
Commit-Ready: Gabe Black <gabeblack@chromium.org>