exynos: video: Fix unintialized variable
The s5p_dp_init_analog_func() function had a local variable called
'timeout'.  This was uninitialized but incremented in the case of a
success (attempt to set it to something non-zero?).  AKA: assuming
that timeout was actually initialized to 0, we'd have had timeout
as 0 in the case of a timeout and 1 in the case of no timeout.

After the loop, if timeout was 0 the debug statement would have
gone off indicating a timeout.

Refactor the code, adding a return on error in the process.

Found by code inspection.

BUG=None
TEST=Compiled.  Tried to exercise this code by using a very small
timeout, but on my current setup I never even make it into
the "== PLL_UNLOCKED" if statement.

Change-Id: I14fcc868ff8907c7cf5eb67d431b5577907100a1
Signed-off-by: Doug Anderson <dianders@chromium.org>
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/25295
Reviewed-by: Simon Glass <sjg@chromium.org>