WIP: Implement basic VbExTrustEC() function.
This is just to get the function in the code. It doesn't yet do the right
thing for the right reasons.

BUG=chrome-os-partner:9953
TEST=none

This must go in with a simultaneous change to vboot_reference. It's only to
get the function into the code. The only test is whether or not everything
continues to build.

Change-Id: I8832ad99ffb915045ec282544a9201781bd7a683
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/25556
Reviewed-by: Randall Spangler <rspangler@chromium.org>