gen: serial: Disable FDT serial console if requested
Rather than ignore the device tree "status" property, use it to determine
whether or not to enable the serial console.

This version leaves the console active, but supresses all output.

BUG=chromium-os:31984
TEST=manual:
Tested on link:
$ cros_bundle_firmware --add-node-enable console <0 or 1>
and see that the console appears (or does not) as requested, and in
any case boots to kernel.

$ cat /sys/firmware/log
See that the output from coreboot and U-Boot still appears regardless of
the console setting.

Change-Id: Idb3b5c814cce56ade7317429cc6e9f25937c44c9
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/25809