x86: Add default console for link
This adds a default console node for link, so that even if coreboot
does not provide console information, we may be able to have a
console.

This means that if coreboot provides a console we will always use it,
and if it doesn't we won't, unless this node is enabled by
cros_bundle_firmware. This way we can have a U-Boot console separately
from the coreboot console.

We may be able to extend this feature to other boards, but will deal
with this as we need to.

BUG=chromium-os:31984
TEST=manual:
Tested on link:
$ emerge-link --nodeps chromeos-bootimage
$ cros_write_firmware -b link -i /build/link/firmware/legacy_image.bin -w em100
Press power button
See that the serial console appears (and also LCD console)

$ cros_write_firmware -b link -i /build/link/firmware/image.bin -w em100
Press power button
See that the serial console does not appear

In both cases:
$ cat /sys/firmware/log
See that the output from coreboot and U-Boot still appears for both
images.

Also tested with cros_bundle_firmware:

$ emerge-link chromeos-coreboot
$ cros_bundle_firmware  -b link -w em100 --add-node-enable console 0
See that U-Boot serial console does not appear

$ cros_bundle_firmware  -b link -w em100 --add-node-enable console 1
See that U-Boot serial console does appear

$ USE=pcserial emerge-link chromeos-coreboot
$ cros_bundle_firmware  -b link -w em100 --add-node-enable console 0
$ cros_bundle_firmware  -b link -w em100 --add-node-enable console 1
See that U-Boot serial console appears in both cases

Change-Id: Ia6d8c1fde2f8bb6da363b02cba883660b9a04f4f
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/25668
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>