gen: Make mkbp more generic, so i2c and spi are optional
Some platforms will not have i2c or spi, so allow this code to be selected
based on new CONFIG_MKBP_SPI and CONFIG_MKBP_I2C options. For now, only
use the GPIO interrupt when either of these is defined.

BUG=chrome-os-parter:10575
TEST=manual:
build and boot on daisy; see that i2c keyboard still works
build and boot on link; see that 8042 keyboard still works
Change-Id: I0fa337f991fdb444349ce57c810ec309d1955ce9
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/25758