gen: Report errors correctly in mkbp_info()
This function can fail, so make it return an error in this case.

BUG=chrome-os-parter:10575
TEST=manual:
build and boot on link without any code to process the message in LPC
mode, and with DEBUG enabled in mkbp.c. See that we now get an error on
init instead of it reporting an empty EC version.
Change-Id: I0b0fd425e19c0e55fd07a385a28ac20967844c2b
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/25760