drop: exynos: video: Add retry for power up of eDP bridge chip
This is a hack and should be replaced by something better when
we figure out what's up with the bridge chip and how we can fix
it.  It also doesn't always work.

On my newer board over 131 boots, I saw 12 transitory failures
that were recovered and then it died totally.  That means that
failure rate before this fix was ~10% and failure rate after
this fix was ~1%.  On a different board failure rate was higher
but still better with this fix than without.

BUG=chrome-os-partner:10259
TEST=Manual:
0. Changed 'debug' statements in s5p_dp_enable_bridge() to printf
   statements and recompiled.
1. Manually changed U-Boot to constantly reboot:
   setenv bootdelay 2
   setenv bootcmd 'echo "rebooting to test eDP"; reset'
   save
2. Reboot for a while.
3. Notice some cases where the retry worked.  Specifically, you'll
   see "s5p_dp_enable_bridge: eDP bridge failed to come up; try 1 of 10"
   On especially bad board, even saw some get to try 2, 3, or even 4
   and then succeed.
4. If you wait long enough, you will see the bridge get totally
   wedged and you'll see all 10 retries happen.  Then you'll see
   an error -17.

Change-Id: I0ab5fd34560dbb8ac32bf1ac58f341baf6def050
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/25805
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>