cros: Do not rescan internal disk and exploit its uniqueness
As the internal disk is not removable, it is simply a waste of time to
rescan it; so do not rescan it.

Besides, as the internal disk is unique, when we are only looking for it
and have just found one, it is safe to stop looking.

These optimizations may save us ~167ms boot time in normal boot path.

Signed-off-by: Che-Liang Chiou <clchiou@chromium.org>

BUG=chrome-os-partner:10660
TEST=On Snow, run normal mode firmware on boot to kernel and see that
     system time on VbExDiskGetInfo reduced from ~333000us to ~166000us,
     which is ~2X speedup.
     Run vbexport_test init && vbexport_test diskinfo
     Also run emerge-x86-alex chromeos-u-boot

Change-Id: I7af35359bddf0327bbe7fc4a7c70933cf841a0cb
Reviewed-on: https://gerrit.chromium.org/gerrit/25830
Tested-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Ready: Che-Liang Chiou <clchiou@chromium.org>