gen: Add i2c arbitration functions
These two new functions are called to claim and release the i2c bus, in a
similar way to how things are done with the SPI bus. If the board implements
these functions, then they will be called by the driver to ensure that
access is available to the I2C bus. This can be used as a way of having
multiple masters on the I2C without needing to use multi-master I2C, which
is complex and sometimes unreliable.

BUG=chrome-os-partner:10906
TEST=cros_test i2c

Change-Id: Ib6c22d8b004cba8e82e8079f77ebfd0976c068a8
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/26212