exynos: Add pull-up on i2c arbitration line from EC
In reality if the EC is not there we are unlikely to be running anyway,
but it is more correct to add a pull-up on the line from the EC, just
in case the EC forgets to drive it.

BUG=chrome-os-partner:10906
TEST=manual:
check that keyboard and battery work ok in the kernel
run 'cros_test i2c' in U-Boot and 'pmu 1000' on EC at the same time.

Change-Id: Iff7f920b7db24d42a072c2757e3ce8fef0c783c3
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/26833