cros: Pass flag from coreboot to vboot regarding VGA Option ROM
For fastest boot, we don't want to load the VGA Option ROM every time, but
only when we need it. Coreboot does that loading, but vboot is what needs to
know whether it happened or not. This CL just passes the "oprom loaded"
flag from coreboot on to vboot.

BUG=chrome-os-partner:8789
TEST=manual

Make sure the system is in normal mode, not dev-mode.

1. Give it the three-finger salute. It should reboot into Recovery mode. At
   the Insert screen, press TAB. You should see "VbSD.flags: 000000e60". The
   important bit is bit 11 ('e'), which should be set.

2. Press Refresh + Power, then Power to boot normally. Log in and run
   "crossystem vdat_flags". You should see "00000e4a". Again, the important
   bit is bit 11 ('e'), which should be set.

There is no other user-visible change.

Change-Id: If4ffebdf0a97bbc0e40832debbde301e4cd9b458
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/26275
Reviewed-by: Randall Spangler <rspangler@chromium.org>