exynos5: check the card detect register of mshci controller
mshci controller has a CDETECT register which hold the information
whether card is persent or not in the physical slot. Controller sense the
card detect (CD) gpio line to know the status of the card. CD line status
change when card is enserted or remove from the slot.
This CL read CDETECT register and return the card status to the mmc
stack.

BUG=chrome-os-partner:10911
TEST= build ok, booted on snow,
	[Remove SD card if any]
	SMDK5250 # mmc dev 1
	MMC: no card present
	mmc1 is current device

	[Now plug in an SD card]

	SMDK5250 # mmc rescan
	ERROR: v7_dcache_inval_range - stop address is not aligned -
	0xbda59f88
	SMDK5250 # mmc read 40008000 0 10

	MMC read: dev # 1, block # 0, count 16 ... 16 blocks read: OK
	SMDK5250 #

Change-Id: I0a821b8504745ebb5100f34eb6776b3c1ec98eb8
Signed-off-by: Alim Akhtar <alim.akhtar@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/26317
Reviewed-by: Simon Glass <sjg@chromium.org>