drop: exynos: Add dummy i2c bus claim/release functions for SPL
Since there is no global i2c core device code, we can't put these
functions in a place which SPL will normally build. In the normal
U-Boot the functions will be in cmd_i2c.c but this file is not built
in SPL mode/

For now, put them here.

BUG=chrome-os-partner:10906
TEST=build U-Boot and see that there are no SPL link errors

Change-Id: I20e732bf5083d9134c323d96dfdea762fef5e51d
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/26344