exynos5: pinumx: Remove pulldown for i2c2
The pins that i2c2 are on are (at bootup) pulled down.  This fights
i2c which should be high in a quiescent state.

We set i2c2 to no pullups since we have a part connected that has
its own pullups and we're slightly out of margins if we have double-
pullups.

This change matches GPIO settings we have in the kernel device tree.
You can see that we've set the 3rd number to 0 (no pull) for both
lines:
  gpios = <&gpa1 2 3 0 0>,
          <&gpa1 3 3 0 0>;

BUG=chrome-os-partner:11039
TEST=Booted before and after and looked at pullup lines in CPU:
Before you see pulldowns (0x5000 = 0101 for top two items):
  SMDK5250 # i2c dev 2
  Setting bus to 2
  SMDK5250 # md.l 11400008 1
  11400008: 00005500    .U..
After you see pullups (0xf000 = 1111 for top two items):
  SMDK5250 # i2c dev 2
  Setting bus to 2
  SMDK5250 # md.l 11400008 1
  11400008: 00000500    ....
TEST=Probed the lines after this change and saw that when the
NXP chip drives the lines low they go a little lower.

Change-Id: I1fb2e02b2aa8956eb2b5b10986779ede3d709bac
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/26585
Reviewed-by: Simon Glass <sjg@chromium.org>