gen: Set kernaddr if fdt indicates a kernel is present
If kernel-offset is specified in the fdt, set an environment variable
so that scripts can access the attached kernel.

This can be used by a packaging program to tell U-Boot about a kernel
that has been downloaded alongside U-Boot. The value in the fdt is
the offset of the kernel from the start of the U-Boot image, so we can
find it just by adding CONFIG_SYS_TEXT_BASE.

It is then fairly easy to put something like this in the environment
variables in the board header file:

		"if test ${kernaddr} != \"\"; then "\
			"echo \"Using bundled kernel\"; "\
			"bootm ${kernaddr};" \
		"fi; "\
		/* rest of boot sequence follows here */

BUG=chromium-os:32386
TEST=manual:
cros_bundle_firmware -b daisy ---kernel /build/daisy/boot/vmlinux.uimg
-d exynos5250-snow
on a snow and see that it boots U-Boot and then picks up the latest
kernel, ignoring the one on the machine.

Change-Id: Ib6135d93f08d24a5c45860aebf5330b45af81019
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/26794
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>