cros: config: Support kernaddr for usb and mmc booting
If this environment variable is set, boot the alternative kernel
instead. This is useful to allow cros_bundle_firmware to download a
kernel to the board for a one-off boot.

BUG=chromium-os:32386
TEST=manual:
cros_bundle_firmware -b daisy ---kernel /build/daisy/boot/vmlinux.uimg
-d exynos5250-snow
on a snow and see that it boots U-Boot and then picks up the latest
kernel, ignoring the one on the machine.

Change-Id: I9c06bc2007856560ab08a22b07dc85bb91dbf60c
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/26795
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>