cros: exynos: Remove TPM burst count cap
The burst count cap of driver's burst count is necessary on Tegra boards
but is not on Exynos boards; so remove it.

While this optimization along does not save us any time, if combined
with the I2C polling optimization, it saves us extra ~261ms.

Signed-off-by: Che-Liang Chiou <clchiou@chromium.org>

BUG=chrome-os-partner:10660
TEST=normal boot on Snow 10 times in a row (for stability testing)

Change-Id: Ic32b924bc743df01ea520521986ce727b2ea9ca8
Reviewed-on: https://gerrit.chromium.org/gerrit/26808
Commit-Ready: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Che-Liang Chiou <clchiou@chromium.org>