gen: ahci: Put all uses of ahci_set_feature() under CONFIG_AHCI_SETFEATURES_XFER
The function ahci_set_feature() is only compiled when the manifest
constant 'CONFIG_AHCI_SETFEATURES_XFER' is defined, but it is used
once outside of such a conditional.

This change puts all uses of the function under the same preprocessor
conditional.

It also fixes a spelling error in a message which can be output to the
console.

BUG=none
TEST=Visual inspection; code compiles.
Signed-off-by: Taylor Hutt <thutt@chromium.org>

Change-Id: I47245d27912edb58eda6896b741879498ac59f53
Reviewed-on: https://gerrit.chromium.org/gerrit/26948
Commit-Ready: Taylor Hutt <thutt@chromium.org>
Reviewed-by: Taylor Hutt <thutt@chromium.org>
Tested-by: Taylor Hutt <thutt@chromium.org>