cros: Pass option ROM input flags to VbInit() as appropriate.
On some systems, we require the VGA option ROM to be loaded before VbInit()
is called so we can display BIOS screens, but may not want it for normal
mode. Other systems may not care.

This passes flags to VbInit() so it can tell whether or not it matters and
what the state is.

BUG=chrome-os-partner:8789
TEST=manual

Using keyboard-based developer mode, switch between normal and dev-mode and
back. It should work as expected.

Change-Id: Ibf2d181932b27330b90166b1b289139d625a7b9f
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27127
Reviewed-by: Randall Spangler <rspangler@chromium.org>